Move StaveSignature into its own file
